moved structs and enums to respective modules
